Server/requirements.txt
PyJWT~=1.7.1
fastapi~=0.62.0
mysql-connector-python~=8.0.24
pydantic~=1.8.1
uvicorn~=0.12.3
requests~=2.25.1
python-decouple~=3.3
numpy~=1.20.2
opencv-python~=4.5.2.54
Pillow~=8.2.0
python-dotenv~=0.17.1